pyproject.toml
[project]
name = "game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.24.0",
    "Pillow>=10.0.0",
    "neo4j>=5.14.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
]
